Hide ‘provide extra information’ preamble when no visible fields are present
For categories that use the extra metadata fields to just show a notice
message to the user, the ‘help the council fix the problem by providing
extra information’ text was a bit out of place. This commit hides it
unless there are actual UI elements also included in the markup.

Fixes #2811.