Default logging to logged in user.
Historically, the FixMyStreet admin was accessed via HTTP Basic Auth and
a secure proxy, so the auth user was used for logging purposes. Nowadays
all admin use is by FixMyStreet user, and any Basic Auth is used for the
purpose of hiding staging sites from bots. This configuration means that
on those staging sites, all admin is logged by the one staging user, not
the actual user, so let's swap that round.