Allow multiple question to disable the new report form
This changes the existing logic to account for the fact that multiple
questions might have the disable form checkbox ticked for one of their
options.

At the moment this just displays one of the messages at a time, with the
first one taking priority. So if you have a "Is this an emergency"
question and a "Is this on private land" question which both disable the
form, and the user has selected "Yes" to both, then only the message for
the first question is displayed. This can potentially be improved in the
future, but seemed out of scope for this change.