[Zürich] Fix crash on certain /report/ajax/<id> URLs
Attempting to load the /report/ajax/<id> URL for a problem that
had a public response and wasn't in the ‘external’ state was causing
a crash because the call to FixMyStreet::App::View::Web::add_links
in updates_as_hashref was returning a FixMyStreet::Template::SafeString
that the JSON module didn't know how to serialise.

This commit adds a TO_JSON method to SafeString, and ensures the output
of /report/ajax is JSON-encoded with convert_blessed turned on so the
TO_JSON method is called.