Fix category extra AJAX call for categories with ampersands
The `IF category_extras.$category.size` condition in category_extras.html
was using an escaped string for `$category`, meaning the lookup would
always fail for categories that had an ampersand in their name. The result
of this was an empty `<div id="category_meta">`, and no extra fields shown
to the user. This caused issues when one or more of the fields were
required, as they were always empty and the error message wasn't being
rendered and shown to the user. This meant the user would see their report
form over and over again when submitting instead of useful feedback.

We also ensure $category is always a SafeString when used as a key into
a hashref in other templates.

Fixes https://github.com/mysociety/fixmystreet-freshdesk/issues/126