[Peterborough] Ensure trees and tree groups share the same config
This fixes a bug where the new tree plantings group was requiring a tree
group to be selected, which isn't the desired behaviour.

The bug was caused by the group layer not having the same `relevant`
function as the points layers. This adds the shared bits into a shared
config that the assets both use.

Also hide the asset item message for request to plant a new tree
category.