[UK] Fix client side HE category submission.
Once a location is fixed, we never want to regenerate the Highways
England question, as that can lose the selected category. Store the
location regardless of the HE question answer, so any category
change, as opposed to a location change, does not do anything.

We also do not want to clear category_meta, as that is where the extra
data fields appear.