Simplify #geolocate_link styling
Much less opinionated styling for #geolocate_link in core, which
means the link should look better, by default, for most cobrands.

(In particular, the link looks much better on cobrands with the more
modern style of white / off-white #front-main background, such as
highwaysengland, which was the cobrand that started off this work
to begin with.)

I’ve also reduced the specificity of the #geolocate_link rulesets
in both core and all the cobrands.

While this commit means that fewer cobrands need to override styles
on #geolocate_link, there are still 8 cobrands that have to define a
custom text colour for their #geolocate_link, because they both:
A) have a light coloured background for #front-main, and
B) want their #geolocate_link to be coloured like a normal link,
rather than inheriting the text colour of the parent element.

We might want to revisit this handling of #geolocate_link colouring and
the `#front-main a { color: inherit }` rule in _layout.scss at some
point in the future.

Fixes mysociety/fixmystreet-commercial#1835.