[Oxfordshire] Add instruct defect functionality back in
In the past Oxfordshire had a way for inspectors to "instruct a defect",
which generated a CSV which got emailed to Oxfordshire with a list of
defects.

They now want to bring this functionality back, but instead of emailing
a CSV they want it to call an API.

As a first step towards that goal, this change adds back some of the
instruct defect functionality that was removed in d94d4ce.