Fix duplicate message after dismissing suggestions
If we create one category_meta_message for each category, removing when
not relevant, then duplicat suggestion showing leads to duplicate
messages because the layer is still relevant, just not visible, and so
the old layer message is not removed. Instead, repurpose the ID of any
existing category_meta_message, so there is only ever one in existence.