Spot user set on update not at object creation.
The changes in 1b91471eb create the necessary metadata at the creation
of a Comment object, but if the user is not included at that point but
later (as is done by the normal public update flow), then it would not
be created. Wrap the user accessor so we can make sure it is set.