Do not set attribute fields on hidden layers.
This did not really matter before because you were either hiding an
asset layer (and thus its fields), or selecting a new layer/asset (and
so the last thing would be setting the attributes for the new asset. But
it was unnecessary, and now we're updating an inspector form and the
fields may apply to more than one layer.