[UK] Pro user guide, link to online version from print version
Spotted by Louise H. URL was broken, and link wasn’t even a link.

Even though this is a print view, it still seems silly not to make the
link work. Added benefit is that if you print to PDF, most browsers
will make the link clickable in the PDF too.