Combine map layer and hide pins toggles in toolbar
A new .map-links element, positioned near pan/zoom controls, replaces
the old .big-hide-pins-link, and acts as a place to store controls to
do with the display of the map. On mobile, the element is only shown
with a full screen map.

As a result, this commit removes the map layer toggle and hide pins
toggles from #sub_map_links.

Zurich already had its own road/aerial toggle, so that’s been
incorporated into the new core functionality (the only difference
being that on Zurich, aerial is the default layer, layer ID 0, while
roads is layer ID 1).