[Highways England] Remove group from categories.
Previously, the fact the Highways England categories were in a group
(for helping with display on .com) did not matter, because groups were
not enabled on the cobrand. But the category work recently enables
groups on all cobrands in the front end, which meant a category/
subcategory arrangement was needlessly being shown.