Make anonymous comments clearer in alert emails
When showing a list of comments in HTML emails we previously didn't show
anything for updates that were left anonymously. This led to some
confusion about which text related to which update.

This change means that anonymous comments will show the text "Posted
anonymously" next to them, which should make things clearer.