Add HTML::Selector::Element to test modules
HTML::Selector::Element makes tests less
brittle, easier to write and more idiomatic
by allowing content and values to be searched
for and specified via the Selectors API.

This allows us to avoid having to use whole
strings of HTML for comparison, which could
break if minor changes are made.

It also makes it much easier to find and check
values in specific areas of the page e.g. to
verify that a particular problem in a list is
showing the right value, when the wrong value
could also be present in another problem listing
also showing which would be caught by
$mech->content_contains().

relates to mysociety/societyworks#2455