[Peterborough] Create new "problem" flow for damaged bins, not "enquiry"
The enquiry flow doesn't fit the way Peterborough want to handle damaged
bin reports. Rather than make that flow too generic it seemed better to
replace it with something closer to what's required.

NB the implementation of this flow is very similar to the existing
"report a missed collection" flow, and it would be a good idea to factor
the two together at some point.