[Bartec] Ensure future collections are shown for new properties
Due to an apparent quirk of the Bartec API, properties that have never
had a collection won't show up in the Jobs_FeatureScheduleDates_Get
response unless we search into the future. In that case the schedules
are returned, but collections that have never happened have a year of
"1900" so we filter those out in the template.

For https://3.basecamp.com/4020879/buckets/19439943/todos/4329332359