[Merton] Give anonymous reports a service
At the moment anonymous reports don't have a service set on them because
the code that sets the service looks at which button was pressed,
desktop or mobile, but doesn't take anonymous into account.

Until that's fixed just give anonymous reports the service of "unknown",
so that when we send the service attribute over open311 (see previous
commit) it has something in there.