[Westminster] Correct "Click map" font size on /around page
Somehow a recent commit applied the same massive `font-size: 3.3em`
to the map page headings, as to the front page heading. This was
particularly unfortunate on the "Click the map…" banner, which is
actually a `h1`.

This commit resets `.mappage` headings (including the "Click the map…"
banner, and the "Report your problem" heading at the top of the
reporting form) to their default font-sizes.

Fixes mysociety/societyworks#2557.