Add prefer_if_multiple setting to contacts
This is for the situation where multiple bodies covering the same area
have a contact with the same name. Checking this box indicates that if
multiple contacts are found with the same name then the one(s) with this
box checked should be preferred. If no contacts are found with this flag
then fall back to the default behaviour of sending to all matching
contacts.