[National Highways] Improve litter handling.
Use the new litter layer to see if National Highways are responsible for
litter on an A-road. On the National Highways site, ask the user to go
to fixmystreet.com if not responsible for litter, otherwise carry on as
normal.

On fixmystreet.com, rework the categories on a non-responsible National
Highways road so that the council litter categories can be selected even
if the National Highways road is chosen.