[Bromley] Update assets JS to handle new sub-categories
Bromley are switching to using "normal" FixMyStreet subcategories, which
means that whereas previously they had categories and the subcategory
was an extra question, now the categories have groups, so they behave
like regular FMS category/subcategory.

This change updates the assets JS to use the new category names and
updates the JS tests to reflect the new way Bromley is setup.