Add Cypress method for picking subcategories
There was already a cy.pickSubcategory method, but the name was
misleading because it was actually picking a subcategory from within
extra data, which isn't how most cobrands work.

This change renames the previous method to pickSubcatExtraInfo to more
accurately reflect what it's doing, then the pickSubcategory method has
been updated to actually pick a regular subcategory given a category ID
and a subcategory label. I've also gone through and updated the obvious
places where this pattern was being used to use the new method.