[Waste] Do not reuse CSC payment method on modify.
If the original payment was made by CSC, this payment method was being
copied across to modifications made by normal users. It is already
re-set to CSC if CSC is used for the modification so default to credit
card in the usual case. Also fix button text for such a payment.