[Kingston] improve pending subscription checks
Because the kingston DD process redirects back to us after submission we
can set some metadata on the new report to indicate this. We can then
use this to check that the pending subscription got as far as submitting
the Direct Debit application, and only indicate there is a pending
subscription if this is the case.