[Waste] Check New records for a dd_date as well.
The script may have created a renewal report the previous day, that
failed to go through first time due to the backend subscription
expiring, it then being changed to a new subscription report which went
through. Then when it comes to the next day, it would warn because the
renewal lacks a property_id to look up (which is lucky because if it had
one, it would have successfully made another renewal report each time).

Fixes https://github.com/mysociety/societyworks/issues/3149