[FMS] Force order of updates if timestamps confusing
An open311 update will be checked to make sure that it is not earlier than the report
it is updating.
It will also be checked to make sure that it is not earlier than an auto-update
on that report.
In either of those cases, it will be amended to be a second later.