[SLWP] Cope with multiple tasks of same type.
As schedules was being indexed by the task type ID, if two service tasks
were of the same type, this could lead to an error looking up the last
instance task as some tasks would not be present in the lookup. Index by
task ID instead, which will be unique.