[Brent] Actions for Housing layer
Brent would like users to visit their website for housing estate issues.

We present a message containing a link to that page if they select a
housing estate.

The housing message stopper requires a criterion to be passed that
returns true to match if there are multiple bodies covering the area.