[FMS] Stop cobrand removing stopper of another on mobile
Adds a class to the mobile alert message that an asset has not been
found based on the id of the alert message. That way a cobrand
will only remove its own messages when triggered through 'not_found'.

Adds a check when enabling the report form to if there are any
mobile asset not found messages.