[Cypress] Try and deal with flaky tests.
One test appears to sometimes not log someone in, so perform the
login more directly. One test appears to sometimes suddenly have
someone logged back in again, meaning the CSRF check fails; this
seems to have something to do with the service worker, though am
not sure what. Disable that while the tests are running. Another
has the same issue (being logged back in again), so there we log
out at the end of the test that logged in (it seems to work more
there than the start of the next test).