Remove subcategory selection on pagehide.
If you click back from an external site to /report/new, we want the
browser not to have remembered your choice of subcategory form field, to
prevent any confusion should that subcategory disable the form or
similar. This will only happen if the bfcache is not involved (in which
case, the right thing will be shown immediately).